Refactor exportable code + change to pixel threshold for segmentation
