+++ external/anomaly/anomaly_classification/exportable_code/__init__.py
@@ -1,0 +1,9 @@
+"""Exportable code for Anomaly Classification task."""
+
+# Copyright (C) 2021-2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from .anomaly_classification import AnomalyClassification
+
+__all__ = ["AnomalyClassification"]
