+++ external/anomaly/anomaly_segmentation/exportable_code/__init__.py
@@ -1,0 +1,9 @@
+"""Exportable code for Anomaly Segmentation task."""
+
+# Copyright (C) 2021-2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from .anomaly_segmentation import AnomalySegmentation
+
+__all__ = ["AnomalySegmentation"]
